Weather forecast tweaks.
Add a custom component to download and parse BOM weather forecasts.
Use that data to let me know the forecast in the morning when the
office is powered up.

In the future, I think I'd like to make smarter decisions about
what to set the office AC to do on startup, but that can be a thing
for another day.